Extract GraphemeHelper to a helper class
